git-svn-id: svn+ssh://ginandtonic.no/home/kms/.svnrepos/scout-kenwood@84 bec0a6a6-59ba-0310-824c-86e0bd1f4ed1

+++ fRound.c
@@ -4,7 +4,6 @@
  * $Id$
  */
 
-#include <stdlib.h>
 #include "fRound.h"
 
 #define NUMBER_OF_BANDS 1
@@ -17,7 +16,7 @@ uint32_t steps[] = {12500};
 
 uint32_t roundFreq(uint32_t freq) {
     uint8_t band;
-    uint16_t r;
+    uint32_t r;
     
     for (band = 1; band < NUMBER_OF_BANDS; band++) {
 	if ((bands[band - 1].lower <= freq)
@@ -27,6 +26,7 @@ uint32_t roundFreq(uint32_t freq) {
     }
     
     band--;
+    
     uint32_t step = steps[bands[band].step];
     r = freq % step;
